scripts/requirements.txt
# 版本管理工具依赖
# 用于 scripts/lib/version_manager.py

PyYAML>=6.0



